requirements-colab.txt
pandas
matplotlib
Pillow
scikit-learn
scikit-image
imgaug
click
tensorflow_datasets